+++ CHANGELOG.rst
@@ -8,7 +8,9 @@ kmpc Change Log
 0.6.1 - 2018-01-??
 ******************
 
-- Added warning to prevent syncing synchost to itself (`issue #73
+- Changed the track slider update task to pause when not on the Now Playing
+  tab. (`issue #64 <https://github.com/eratosthene/kmpc/issues/64`_)
+- Added warning to prevent syncing synchost to itself. (`issue #73
   <https://github.com/eratosthene/kmpc/issues/73>`_)
 - Changed sections in config file and updated docs accordingly to better
   account for which fields are used in each program. (`issue #70